Revert "Use 64 bit parameter Id"
This reverts commit 2a9ad226fab80135dc2f191097e319ef3685ee5f.

+++ components/cAVS/include/cAVS/DspFw/Common.hpp
@@ -46,7 +46,7 @@ namespace detail
 /** Unique trait to for parameterId. */
 struct ParameterIdTrait
 {
-    using RawType = uint64_t;
+    using RawType = uint32_t;
 };
 
 } // namespace detail
